+++ bindings/golang/uuid7.go
@@ -50,6 +50,14 @@ func (g *Generator) Close() {
 	}
 }
 
+// UUIDToString memformat UUID ke representasi string standar RFC 4122:
+// xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
 func UUIDToString(uuid [16]byte) string {
-	return fmt.Sprintf("%x", uuid)
+	return fmt.Sprintf("%x-%x-%x-%x-%x",
+		uuid[0:4],
+		uuid[4:6],
+		uuid[6:8],
+		uuid[8:10],
+		uuid[10:16],
+	)
 }